<- Click on the number here if you read on Slack to have all the commit message
API :
o Modification of the filter of music : file is not given anymore
o The get function in music controller of the API is now not deprecated. There is only one music on the repository now : the id 1. The structure of the folder musics is : /musics/artist_id/music_file.mp3 (or cut_music_file.mp3). The file is set in the database.

Front :
o Responsiveness of the player button
o Work on the playlist management